+++ butane_core/tests/migration.rs
@@ -1,764 +1,580 @@
-use butane_core::db::ConnectionAsync;
-use butane_core::migrations::adb::*;
-use butane_core::SqlType;
-use butane_test_helper::*;
-use butane_test_macros::butane_test;
+use butane_core::codegen::{butane_type_with_migrations, model_with_migrations};
+use butane_core::db::{BackendConnection, Connection};
+use butane_core::migrations::adb::{DeferredSqlType, TypeIdentifier, TypeKey};
+use butane_core::migrations::{MemMigrations, Migration, MigrationMut, Migrations, MigrationsMut};
+use butane_core::{SqlType, SqlVal};
+#[cfg(feature = "pg")]
+use butane_test_helper::pg_connection;
+#[cfg(feature = "sqlite")]
+use butane_test_helper::sqlite_connection;
+use proc_macro2::TokenStream;
+use quote::quote;
+use sqlparser::dialect::GenericDialect;
+use sqlparser::parser::Parser as SqlParser;
 
 #[test]
-fn empty_diff() {
-    let old = ADB::default();
-    let new = ADB::default();
-    let ops = diff(&old, &new);
-    assert_eq!(ops, vec![]);
+fn current_migration_basic() {
+    let tokens = quote! {
+        struct Foo {
+            id: i64,
+            bar: String,
+            baz: f64,
+        }
+    };
+
+    let mut ms = MemMigrations::new();
+    model_with_migrations(tokens, &mut ms);
+    let m = ms.current();
+    let db = m.db().unwrap();
+    assert_eq!(db.tables().count(), 1);
+    let table = db.get_table("Foo").expect("No Foo table");
+    let idcol = table.column("id").unwrap();
+    assert_eq!(idcol.name(), "id");
+    assert!(!idcol.nullable());
+    assert!(idcol.is_pk());
+    assert_eq!(*idcol.default(), None);
+    assert_eq!(idcol.typeid().unwrap(), TypeIdentifier::Ty(SqlType::BigInt));
+    assert!(!idcol.is_auto());
+
+    let barcol = table.column("bar").unwrap();
+    assert_eq!(barcol.name(), "bar");
+    assert!(!barcol.nullable());
+    assert!(!barcol.is_pk());
+    assert_eq!(*barcol.default(), None);
+    assert_eq!(barcol.typeid().unwrap(), TypeIdentifier::Ty(SqlType::Text));
+    assert!(!barcol.is_auto());
+
+    let baz_col = table.column("baz").unwrap();
+    assert_eq!(baz_col.name(), "baz");
+    assert!(!baz_col.nullable());
+    assert!(!baz_col.is_pk());
+    assert_eq!(*baz_col.default(), None);
+    assert_eq!(baz_col.typeid().unwrap(), TypeIdentifier::Ty(SqlType::Real));
+    assert!(!baz_col.is_auto());
+
+    assert_eq!(table.pk(), Some(idcol))
 }
 
 #[test]
-fn add_table() {
-    let old = ADB::default();
-    let mut new = ADB::default();
-    let mut table = ATable::new("a".to_owned());
-    let column = AColumn::new_simple(
-        "a".to_owned(),
-        DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)),
-    );
-    table.add_column(column);
-    new.replace_table(table.clone());
+fn current_migration_pk_attribute() {
+    let tokens = quote! {
+        #[derive(PartialEq, Eq, Debug, Clone)]
+        struct Foo {
+            #[pk]
+            name: String,
+            bar: String,
+        }
+    };
 
-    let ops = diff(&old, &new);
+    let mut ms = MemMigrations::new();
+    model_with_migrations(tokens, &mut ms);
+    let m = ms.current();
+    let db = m.db().unwrap();
+    let table = db.get_table("Foo").expect("No Foo table");
+    let pkcol = table.column("name").unwrap();
+    assert!(pkcol.is_pk());
 
-    assert_eq!(ops, vec![Operation::AddTable(table)]);
+    assert_eq!(table.pk(), Some(pkcol))
 }
 
 #[test]
-fn remove_table() {
-    let mut old = ADB::default();
-    let new = ADB::default();
-    let mut table = ATable::new("a".to_owned());
-    let column = AColumn::new_simple(
-        "a".to_owned(),
-        DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)),
-    );
-    table.add_column(column);
-    old.replace_table(table.clone());
+fn current_migration_default_attribute() {
+    let tokens = quote! {
+        #[derive(PartialEq, Eq, Debug, Clone)]
+        struct Foo {
+            id: i64,
+            #[default="turtle"]
+            bar: String,
+        }
+    };
 
-    let ops = diff(&old, &new);
-
-    let expected_ops = vec![
-        Operation::RemoveTableConstraints(table),
-        Operation::RemoveTable("a".to_owned()),
-    ];
-    assert_eq!(ops, expected_ops);
+    let mut ms = MemMigrations::new();
+    model_with_migrations(tokens, &mut ms);
+    let m = ms.current();
+    let db = m.db().unwrap();
+    let table = db.get_table("Foo").expect("No Foo table");
+    let barcol = table.column("bar").unwrap();
+    assert_eq!(*barcol.default(), Some(SqlVal::Text("turtle".to_string())));
 }
 
 #[test]
-fn stable_table_alpha_order() {
-    let old = ADB::default();
-    let mut new = ADB::default();
+fn current_migration_auto_attribute() {
+    let tokens = quote! {
+        #[derive(PartialEq, Eq, Debug, Clone)]
+        struct Foo {
+            id: AutoPk<i64>,
+            bar: String,
+        }
+    };
 
-    // Insert tables out of order
-    let mut table_b = ATable::new("b".to_owned());
-    let column_b = AColumn::new_simple(
-        "b".to_owned(),
-        DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)),
-    );
-    table_b.add_column(column_b);
-    new.replace_table(table_b.clone());
+    let mut ms = MemMigrations::new();
+    model_with_migrations(tokens, &mut ms);
+    let m = ms.current();
+    let db = m.db().unwrap();
+    let table = db.get_table("Foo").expect("No Foo table");
+    let idcol = table.column("id").unwrap();
+    assert!(idcol.is_auto());
+}
 
-    let mut table_a = ATable::new("a".to_owned());
-    let column_a = AColumn::new_simple(
-        "a".to_owned(),
-        DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)),
-    );
-    table_a.add_column(column_a);
-    new.replace_table(table_a.clone());
+#[test]
+fn current_migration_nullable_col() {
+    let tokens = quote! {
+        #[derive(PartialEq, Eq, Debug, Clone)]
+        struct Foo {
+            id: i64,
+            bar: Option<String>,
+        }
+    };
 
-    let ops = diff(&old, &new);
+    let mut ms = MemMigrations::new();
+    model_with_migrations(tokens, &mut ms);
+    let m = ms.current();
+    let db = m.db().unwrap();
+    let table = db.get_table("Foo").expect("No Foo table");
+    let col = table.column("bar").unwrap();
+    assert!(col.nullable());
+    assert_eq!(col.typeid().unwrap(), TypeIdentifier::Ty(SqlType::Text));
+}
 
+#[test]
+fn current_migration_custom_type() {
+    let tokens = quote! {
+        #[derive(PartialEq, Eq, Debug, Clone)]
+        enum Frobnozzle {
+            Foo,
+            Bar,
+            Baz,
+        }
+    };
+    let mut ms = MemMigrations::new();
+    butane_type_with_migrations(quote! {Text}, tokens, &mut ms);
+
+    let tokens = quote! {
+        #[derive(PartialEq, Eq, Debug, Clone)]
+        struct HasCustomField {
+            id: i64,
+            frob: Frobnozzle,
+        }
+    };
+    model_with_migrations(tokens, &mut ms);
+
+    let m = ms.current();
+    let db = m.db().unwrap();
+    eprintln!("types {:?}", db.types());
     assert_eq!(
-        ops,
-        vec![Operation::AddTable(table_a), Operation::AddTable(table_b)]
+        db.types()
+            .get(&TypeKey::CustomType("Frobnozzle".to_string())),
+        Some(&DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)))
     );
+    let table = db
+        .get_table("HasCustomField")
+        .expect("No HasCustomField table");
+    let col = table.column("frob").expect("No frob field");
+    assert_eq!(col.typeid().unwrap(), TypeIdentifier::Ty(SqlType::Text));
 }
 
+#[cfg(feature = "sqlite")]
 #[test]
-fn stable_new_table_column_insert_order() {
-    let old = ADB::default();
-    let mut new = ADB::default();
-    let mut table = ATable::new("a".to_owned());
-
-    // Insert columns out of order
-    let column_b = AColumn::new_simple(
-        "b".to_owned(),
-        DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)),
+fn migration_add_field_sqlite() {
+    migration_add_field(
+        &mut sqlite_connection(),
+        "ALTER TABLE Foo ADD COLUMN baz INTEGER NOT NULL DEFAULT 0;",
+        // The exact details of futzing a DROP COLUMN in sqlite aren't
+        // important (e.g. the temp table naming is certainly not part
+        // of the API contract), but the goal here is to ensure we're
+        // getting sane looking downgrade sql and a test failure if it
+        // changes. If the change is innocuous, this test should just
+        // be updated.
+        "CREATE TABLE Foo__butane_tmp (id INTEGER NOT NULL PRIMARY KEY,bar TEXT NOT NULL);
+INSERT INTO Foo__butane_tmp SELECT id, bar FROM Foo;DROP TABLE Foo;
+ALTER TABLE Foo__butane_tmp RENAME TO Foo;",
     );
-    table.add_column(column_b.clone());
-
-    let column_a = AColumn::new_simple(
-        "a".to_owned(),
-        DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)),
-    );
-    table.add_column(column_a.clone());
-
-    // Add updated table to new
-    new.replace_table(table.clone());
-
-    let ops = diff(&old, &new);
-
-    // Columns remain in insertion order
-    assert_eq!(ops, vec![Operation::AddTable(table)]);
 }
 
+#[cfg(feature = "pg")]
 #[test]
-fn stable_add_column_alpha_order() {
-    let mut old = ADB::default();
-    let mut new = ADB::default();
-    let mut table = ATable::new("a".to_owned());
-
-    // Add empty table to old
-    old.replace_table(table.clone());
-
-    // Insert columns out of order
-    let column_b = AColumn::new_simple(
-        "b".to_owned(),
-        DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)),
-    );
-    table.add_column(column_b.clone());
-
-    let column_a = AColumn::new_simple(
-        "a".to_owned(),
-        DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Text)),
-    );
-    table.add_column(column_a.clone());
-
-    // Add updated table to new
-    new.replace_table(table.clone());
-
-    let ops = diff(&old, &new);
-
-    assert_eq!(
-        ops,
-        vec![
-            Operation::AddColumn("a".to_owned(), column_a),
-            Operation::AddColumn("a".to_owned(), column_b),
-        ]
+fn migration_add_field_pg() {
+    let (mut conn, _data) = pg_connection();
+    migration_add_field(
+        &mut conn,
+        "ALTER TABLE Foo ADD COLUMN baz BIGINT NOT NULL DEFAULT 0;",
+        "ALTER TABLE Foo DROP COLUMN baz;",
     );
 }
 
+#[cfg(feature = "sqlite")]
 #[test]
-fn add_table_fkey() {
-    let known_int_type = DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Int));
-
-    let old = ADB::default();
-    let mut new = ADB::default();
-    let mut table_a = ATable::new("a".to_owned());
-    let column = AColumn::new(
-        "id".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        None,  // reference
-    );
-    table_a.add_column(column);
-    new.replace_table(table_a.clone());
-
-    let mut table_b = ATable::new("b".to_owned());
-    let column = AColumn::new(
-        "fkey".to_owned(),
-        DeferredSqlType::Deferred(TypeKey::PK("a".to_owned())),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Deferred(DeferredSqlType::Deferred(TypeKey::PK(
-            "a".to_owned(),
-        )))),
-    );
-    table_b.add_column(column);
-    new.replace_table(table_b.clone());
-
-    new.resolve_types().unwrap();
-
-    let mut resolved_table_b = ATable::new("b".to_owned());
-    let column = AColumn::new(
-        "fkey".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Literal(ARefLiteral::new(
-            "a".to_owned(),
-            "id".to_owned(),
-        ))),
-    );
-    resolved_table_b.add_column(column);
-
-    let ops = diff(&old, &new);
-
-    assert_eq!(
-        ops,
-        vec![
-            Operation::AddTable(table_a),
-            Operation::AddTable(resolved_table_b.clone()),
-            Operation::AddTableConstraints(resolved_table_b),
-        ]
+fn migration_add_field_with_default_sqlite() {
+    migration_add_field_with_default(
+        &mut sqlite_connection(),
+        "ALTER TABLE Foo ADD COLUMN baz INTEGER NOT NULL DEFAULT 42;",
+        // See comments on migration_add_field_sqlite
+        r#"CREATE TABLE Foo__butane_tmp (id INTEGER NOT NULL PRIMARY KEY,bar TEXT NOT NULL);
+            INSERT INTO Foo__butane_tmp SELECT id, bar FROM Foo;
+            DROP TABLE Foo;ALTER TABLE Foo__butane_tmp RENAME TO Foo;"#,
     );
 }
 
-/// This is the same as test "add_table_fkey", except that it
-/// runs the DDL on a database, and then deletes the column.
-#[butane_test(nomigrate)]
-async fn add_table_fkey_delete_column(conn: ConnectionAsync) {
-    let known_int_type = DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Int));
-
-    let old = ADB::default();
-    let mut new = ADB::default();
-    let mut table_a = ATable::new("a".to_owned());
-    let id_column = AColumn::new(
-        "id".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        None,  // reference
+#[cfg(feature = "pg")]
+#[test]
+fn migration_add_field_with_default_pg() {
+    let (mut conn, _data) = pg_connection();
+    migration_add_field_with_default(
+        &mut conn,
+        "ALTER TABLE Foo ADD COLUMN baz BIGINT NOT NULL DEFAULT 42;",
+        "ALTER TABLE Foo DROP COLUMN baz;",
     );
-    table_a.add_column(id_column.clone());
-    new.replace_table(table_a.clone());
+}
 
-    let mut table_b = ATable::new("b".to_owned());
-    table_b.add_column(id_column.clone());
-    let column = AColumn::new(
-        "fkey".to_owned(),
-        DeferredSqlType::Deferred(TypeKey::PK("a".to_owned())),
-        false, // nullable
-        false, // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Deferred(DeferredSqlType::Deferred(TypeKey::PK(
-            "a".to_owned(),
-        )))),
+#[cfg(feature = "pg")]
+#[test]
+fn migration_modify_field_pg() {
+    env_logger::try_init().ok();
+    let (mut conn, _data) = pg_connection();
+    // Not verifying rename right now because we don't detect it
+    // https://github.com/Electron100/butane/issues/89
+
+    migration_modify_field_type_change(
+        &mut conn,
+        "ALTER TABLE Foo ALTER COLUMN bar SET DATA TYPE BIGINT;",
+        "ALTER TABLE Foo ALTER COLUMN bar SET DATA TYPE INTEGER;",
     );
-    table_b.add_column(column);
-    new.replace_table(table_b.clone());
 
-    new.resolve_types().unwrap();
-
-    let mut resolved_table_b = ATable::new("b".to_owned());
-    resolved_table_b.add_column(id_column);
-    let column = AColumn::new(
-        "fkey".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        false, // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Literal(ARefLiteral::new(
-            "a".to_owned(),
-            "id".to_owned(),
-        ))),
+    migration_modify_field_nullability_change(
+        &mut conn,
+        "ALTER TABLE Foo ALTER COLUMN bar DROP NOT NULL;",
+        "ALTER TABLE Foo ALTER COLUMN bar SET NOT NULL;",
     );
-    resolved_table_b.add_column(column);
 
-    let ops = diff(&old, &new);
-
-    assert_eq!(
-        ops,
-        vec![
-            Operation::AddTable(table_a),
-            Operation::AddTable(resolved_table_b.clone()),
-            Operation::AddTableConstraints(resolved_table_b),
-        ]
+    migration_modify_field_pkey_change(
+        &mut conn,
+        "ALTER TABLE Foo DROP CONSTRAINT IF EXISTS Foo_pkey;\nALTER TABLE Foo ADD PRIMARY KEY (baz);",
+        "ALTER TABLE Foo DROP CONSTRAINT IF EXISTS Foo_pkey;\nALTER TABLE Foo ADD PRIMARY KEY (bar);",
     );
+
+    migration_modify_field_uniqueness_change(
+        &mut conn,
+        "ALTER TABLE Foo ADD UNIQUE (bar);",
+        "ALTER TABLE Foo DROP CONSTRAINT Foo_bar_key;",
+    );
+
+    migration_modify_field_default_added(
+        &mut conn,
+        "ALTER TABLE Foo ALTER COLUMN bar SET DEFAULT 42;",
+        "ALTER TABLE Foo ALTER COLUMN bar DROP DEFAULT;",
+    );
+
+    migration_modify_field_different_default(
+        &mut conn,
+        "ALTER TABLE Foo ALTER COLUMN bar SET DEFAULT 42;",
+        "ALTER TABLE Foo ALTER COLUMN bar SET DEFAULT 41;",
+    );
+}
+
+#[cfg(feature = "sqlite")]
+#[test]
+fn migration_add_and_remove_field_sqlite() {
+    migration_add_and_remove_field(
+        &mut sqlite_connection(),
+        // The exact details of futzing a DROP COLUMN in sqlite aren't
+        // important (e.g. the temp table naming is certainly not part
+        // of the API contract), but the goal here is to ensure we're
+        // getting sane looking downgrade sql and a test failure if it
+        // changes. If the change is innocuous, this test should just
+        // be updated.
+        r#"ALTER TABLE Foo ADD COLUMN baz INTEGER NOT NULL DEFAULT 0;
+            CREATE TABLE Foo__butane_tmp (id INTEGER NOT NULL PRIMARY KEY,baz INTEGER NOT NULL);
+            INSERT INTO Foo__butane_tmp SELECT id, baz FROM Foo;
+            DROP TABLE Foo;ALTER TABLE Foo__butane_tmp RENAME TO Foo;"#,
+        r#"ALTER TABLE Foo ADD COLUMN bar TEXT NOT NULL DEFAULT '';
+            CREATE TABLE Foo__butane_tmp (id INTEGER NOT NULL PRIMARY KEY,bar TEXT NOT NULL);
+            INSERT INTO Foo__butane_tmp SELECT id, bar FROM Foo;DROP TABLE Foo;
+            ALTER TABLE Foo__butane_tmp RENAME TO Foo;"#,
+    );
+}
+
+#[cfg(feature = "pg")]
+#[test]
+fn migration_add_and_remove_field_pg() {
+    let (mut conn, _data) = pg_connection();
+    migration_add_and_remove_field(
+        &mut conn,
+        "ALTER TABLE Foo ADD COLUMN baz BIGINT NOT NULL DEFAULT 0;ALTER TABLE Foo DROP COLUMN bar;",
+        "ALTER TABLE Foo ADD COLUMN bar TEXT NOT NULL DEFAULT '';ALTER TABLE Foo DROP COLUMN baz;",
+    );
+}
+
+#[cfg(feature = "sqlite")]
+#[test]
+fn migration_delete_table_sqlite() {
+    migration_delete_table(
+        &mut sqlite_connection(),
+        "DROP TABLE Foo;",
+        "CREATE TABLE Foo (id INTEGER NOT NULL PRIMARY KEY,bar TEXT NOT NULL);",
+    );
+}
+
+#[cfg(feature = "pg")]
+#[test]
+fn migration_delete_table_pg() {
+    let (mut conn, _data) = pg_connection();
+    migration_delete_table(
+        &mut conn,
+        "DROP TABLE Foo;",
+        "CREATE TABLE Foo (id BIGINT NOT NULL PRIMARY KEY,bar TEXT NOT NULL);",
+    );
+}
+
+fn test_migrate(
+    conn: &mut Connection,
+    init_tokens: TokenStream,
+    v2_tokens: TokenStream,
+    expected_up_sql: &str,
+    expected_down_sql: &str,
+) {
+    let mut ms = MemMigrations::new();
+    let backend = conn.backend();
+    let backends = nonempty::nonempty![backend];
+    model_with_migrations(init_tokens, &mut ms);
+    assert!(ms.create_migration(&backends, "init", None).unwrap());
+
+    model_with_migrations(v2_tokens, &mut ms);
+    assert!(ms
+        .create_migration(&backends, "v2", ms.latest().as_ref())
+        .unwrap());
+
+    let to_apply = ms.unapplied_migrations(conn).unwrap();
+    assert_eq!(to_apply.len(), 2);
+
+    ms.migrate(conn).unwrap();
+
+    let to_apply = ms.unapplied_migrations(conn).unwrap();
+    assert_eq!(to_apply.len(), 0);
+
+    verify_sql(conn, &ms, expected_up_sql, expected_down_sql);
+
+    // Now downgrade, just to make sure we can
+    ms.unmigrate(conn).unwrap();
+
+    let to_apply = ms.unapplied_migrations(conn).unwrap();
+    assert_eq!(to_apply.len(), 2);
+}
+
+fn verify_sql(
+    conn: &Connection,
+    ms: &impl Migrations,
+    expected_up_sql: &str,
+    expected_down_sql: &str,
+) {
+    let dialect = GenericDialect {};
+    let expected_up_ast = SqlParser::parse_sql(&dialect, expected_up_sql).unwrap();
+    let expected_down_ast = SqlParser::parse_sql(&dialect, expected_down_sql).unwrap();
 
     let backend = conn.backend();
-    let sql = backend.create_migration_sql(&new, ops).unwrap();
+    let v2_migration = ms.latest().unwrap();
 
-    conn.execute(&sql).await.unwrap();
-    conn.execute("SELECT * from a").await.unwrap();
-    conn.execute("SELECT * from b").await.unwrap();
-
-    // "ALTER TABLE b DROP COLUMN fkey;" fails due to sqlite not being
-    // able to remove the attached constraint, however the RemoveColumn
-    // operation already recreates the table, so this works.
-    let remove_column_op = Operation::RemoveColumn("b".to_owned(), "fkey".to_owned());
-    let sql = backend
-        .create_migration_sql(&new, vec![remove_column_op])
-        .unwrap();
-    conn.execute(&sql).await.unwrap();
+    let actual_up_sql = v2_migration.up_sql(backend.name()).unwrap().unwrap();
+    let actual_up_ast = sqlparser::parser::Parser::parse_sql(&dialect, &actual_up_sql).unwrap();
+    assert_eq!(actual_up_ast, expected_up_ast);
+    let actual_down_sql = v2_migration.down_sql(backend.name()).unwrap().unwrap();
+    let actual_down_ast = sqlparser::parser::Parser::parse_sql(&dialect, &actual_down_sql).unwrap();
+    assert_eq!(actual_down_ast, expected_down_ast);
 }
 
-/// This is the same as test "add_table_fkey", except that it
-/// intentionally links a column on table a to table b, and
-/// it runs the DDL on a database.
-#[butane_test(nomigrate)]
-async fn add_table_fkey_back_reference(conn: ConnectionAsync) {
-    let known_int_type = DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Int));
+fn migration_add_field(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            id: i64,
+            bar: String,
+        }
+    };
 
-    let old = ADB::default();
-    let mut new = ADB::default();
-    let mut table_b = ATable::new("b".to_owned());
-    let column = AColumn::new(
-        "id".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        None,  // reference
-    );
-    table_b.add_column(column);
-    new.replace_table(table_b.clone());
+    let v2 = quote! {
+        struct Foo {
+            id: i64,
+            bar: String,
+            baz: u32,
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
 
-    let mut table_a = ATable::new("a".to_owned());
-    let column = AColumn::new(
-        "fkey".to_owned(),
-        DeferredSqlType::Deferred(TypeKey::PK("b".to_owned())),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Deferred(DeferredSqlType::Deferred(TypeKey::PK(
-            "b".to_owned(),
-        )))),
-    );
-    table_a.add_column(column);
-    new.replace_table(table_a.clone());
+fn migration_add_field_with_default(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            id: i64,
+            bar: String,
+        }
+    };
 
-    new.resolve_types().unwrap();
+    let v2 = quote! {
+        struct Foo {
+            id: i64,
+            bar: String,
+            #[default=42]
+            baz: u32,
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
 
-    let mut resolved_table_a = ATable::new("a".to_owned());
-    let column = AColumn::new(
-        "fkey".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Literal(ARefLiteral::new(
-            "b".to_owned(),
-            "id".to_owned(),
-        ))),
-    );
-    resolved_table_a.add_column(column);
+fn migration_modify_field_type_change(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            id: i64,
+            bar: i32,
+        }
+    };
 
-    let ops = diff(&old, &new);
+    let v2 = quote! {
+        struct Foo {
+            id: i64,
+            bar: i64,
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
 
-    assert_eq!(
-        ops,
-        vec![
-            Operation::AddTable(resolved_table_a.clone()),
-            Operation::AddTable(table_b),
-            Operation::AddTableConstraints(resolved_table_a),
-        ]
-    );
+fn migration_modify_field_nullability_change(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            id: i64,
+            bar: i32,
+        }
+    };
 
+    let v2 = quote! {
+        struct Foo {
+            id: i64,
+            bar: Option<i32>,
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
+
+fn migration_modify_field_uniqueness_change(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            id: i64,
+            bar: i32,
+        }
+    };
+
+    let v2 = quote! {
+        struct Foo {
+            id: i64,
+            #[unique]
+            bar: i32,
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
+
+fn migration_modify_field_pkey_change(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            #[pk]
+            bar: i64,
+            baz: i32,
+        }
+    };
+
+    let v2 = quote! {
+        struct Foo {
+            bar: i64,
+            #[pk]
+            baz: i32
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
+
+fn migration_modify_field_default_added(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            id: i64,
+            bar: String,
+        }
+    };
+
+    let v2 = quote! {
+        struct Foo {
+            id: i64,
+            #[default=42]
+            bar: String,
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
+
+fn migration_modify_field_different_default(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            id: i64,
+            #[default=41]
+            bar: String,
+        }
+    };
+
+    let v2 = quote! {
+        struct Foo {
+            id: i64,
+            #[default=42]
+            bar: String,
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
+
+fn migration_add_and_remove_field(conn: &mut Connection, up_sql: &str, down_sql: &str) {
+    let init = quote! {
+        struct Foo {
+            id: i64,
+            bar: String,
+        }
+    };
+
+    let v2 = quote! {
+        struct Foo {
+            id: i64,
+            baz: u32,
+        }
+    };
+    test_migrate(conn, init, v2, up_sql, down_sql);
+}
+
+fn migration_delete_table(conn: &mut Connection, expected_up_sql: &str, expected_down_sql: &str) {
+    let init_tokens = quote! {
+        struct Foo {
+            id: i64,
+            bar: String,
+        }
+    };
+
+    let mut ms = MemMigrations::new();
     let backend = conn.backend();
-    let sql = backend.create_migration_sql(&new, ops).unwrap();
-    let sql_lines: Vec<&str> = sql.lines().collect();
-    if backend.name() == "sqlite" {
-        assert_eq!(
-            sql_lines,
-            vec![
-                "CREATE TABLE a (",
-                "fkey INTEGER NOT NULL PRIMARY KEY,",
-                "FOREIGN KEY (fkey) REFERENCES b(id)",
-                ");",
-                "CREATE TABLE b (",
-                "id INTEGER NOT NULL PRIMARY KEY",
-                ");",
-            ]
-        );
-    }
+    let backends = nonempty::nonempty![backend];
+    model_with_migrations(init_tokens, &mut ms);
+    assert!(ms.create_migration(&backends, "init", None).unwrap());
 
-    conn.execute(&sql).await.unwrap();
-    conn.execute("SELECT * from a").await.unwrap();
-    conn.execute("SELECT * from b").await.unwrap();
-}
+    ms.current().delete_table("Foo").unwrap();
+    assert!(ms
+        .create_migration(&backends, "v2", ms.latest().as_ref())
+        .unwrap());
 
-/// This is the same as test "add_table_fkey", except that it
-/// creates a table with multiple fkey constraints.
-#[butane_test(nomigrate)]
-async fn add_table_fkey_multiple(conn: ConnectionAsync) {
-    let known_int_type = DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Int));
+    let to_apply = ms.unapplied_migrations(conn).unwrap();
+    assert_eq!(to_apply.len(), 2);
 
-    let old = ADB::default();
-    let mut new = ADB::default();
+    ms.migrate(conn).unwrap();
 
-    let id_column = AColumn::new(
-        "id".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        None,  // reference
-    );
+    let to_apply = ms.unapplied_migrations(conn).unwrap();
+    assert_eq!(to_apply.len(), 0);
 
-    let mut table_a = ATable::new("a".to_owned());
-    table_a.add_column(id_column.clone());
-    new.replace_table(table_a.clone());
+    verify_sql(conn, &ms, expected_up_sql, expected_down_sql);
 
-    let mut table_b = ATable::new("b".to_owned());
-    table_b.add_column(id_column.clone());
-    new.replace_table(table_b.clone());
+    // Now downgrade, just to make sure we can
+    ms.unmigrate(conn).unwrap();
 
-    let mut table_c = ATable::new("c".to_owned());
-    table_c.add_column(id_column.clone());
-    let column = AColumn::new(
-        "fkey_a".to_owned(),
-        DeferredSqlType::Deferred(TypeKey::PK("a".to_owned())),
-        false, // nullable
-        false, // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Deferred(DeferredSqlType::Deferred(TypeKey::PK(
-            "a".to_owned(),
-        )))),
-    );
-    table_c.add_column(column);
-    let column = AColumn::new(
-        "fkey_b".to_owned(),
-        DeferredSqlType::Deferred(TypeKey::PK("b".to_owned())),
-        false, // nullable
-        false, // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Deferred(DeferredSqlType::Deferred(TypeKey::PK(
-            "b".to_owned(),
-        )))),
-    );
-    table_c.add_column(column);
-    new.replace_table(table_c.clone());
-
-    new.resolve_types().unwrap();
-
-    let mut resolved_table_c = ATable::new("c".to_owned());
-    resolved_table_c.add_column(id_column);
-    let column = AColumn::new(
-        "fkey_a".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        false, // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Literal(ARefLiteral::new(
-            "a".to_owned(),
-            "id".to_owned(),
-        ))),
-    );
-    resolved_table_c.add_column(column);
-    let column = AColumn::new(
-        "fkey_b".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        false, // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Literal(ARefLiteral::new(
-            "b".to_owned(),
-            "id".to_owned(),
-        ))),
-    );
-    resolved_table_c.add_column(column);
-
-    let ops = diff(&old, &new);
-
-    assert_eq!(
-        ops,
-        vec![
-            Operation::AddTable(table_a),
-            Operation::AddTable(table_b),
-            Operation::AddTable(resolved_table_c.clone()),
-            Operation::AddTableConstraints(resolved_table_c),
-        ]
-    );
-
-    let backend = conn.backend();
-    let sql = backend.create_migration_sql(&new, ops).unwrap();
-
-    conn.execute(&sql).await.unwrap();
-    conn.execute("SELECT * from a").await.unwrap();
-    conn.execute("SELECT * from b").await.unwrap();
-}
-
-/// Creates the test case for adding a foreign key, returning the migration operations,
-/// the target ADB, and the tables which should be expected to be created.
-fn create_add_renamed_table_fkey_ops() -> (Vec<Operation>, ADB, ATable, ATable) {
-    let known_int_type = DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Int));
-
-    let old = ADB::default();
-    let mut new = ADB::default();
-    let mut table_a = ATable::new("a_table".to_owned());
-    let column = AColumn::new(
-        "id".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        None,  // reference
-    );
-    table_a.add_column(column);
-    new.replace_table(table_a.clone());
-
-    // Add the type "a" renamed to table "a_table"
-    let type_name = TypeKey::PK("a".to_owned());
-    let table_name = DeferredSqlType::Deferred(TypeKey::PK("a_table".to_owned()));
-    new.add_type(type_name, table_name);
-
-    let mut table_b = ATable::new("b".to_owned());
-    let column = AColumn::new(
-        "b".to_owned(),
-        DeferredSqlType::Deferred(TypeKey::PK("a".to_owned())),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Deferred(DeferredSqlType::Deferred(TypeKey::PK(
-            "a".to_owned(),
-        )))),
-    );
-    table_b.add_column(column);
-    new.replace_table(table_b.clone());
-
-    new.resolve_types().unwrap();
-
-    let mut resolved_table_b = ATable::new("b".to_owned());
-    let column = AColumn::new(
-        "b".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Literal(ARefLiteral::new(
-            "a_table".to_owned(),
-            "id".to_owned(),
-        ))),
-    );
-    resolved_table_b.add_column(column);
-
-    let ops = diff(&old, &new);
-
-    (ops, new, table_a, resolved_table_b)
-}
-
-#[test]
-fn add_renamed_table_fkey() {
-    let (ops, _, table_a, resolved_table_b) = create_add_renamed_table_fkey_ops();
-
-    assert_eq!(
-        ops,
-        vec![
-            Operation::AddTable(table_a),
-            Operation::AddTable(resolved_table_b.clone()),
-            Operation::AddTableConstraints(resolved_table_b),
-        ]
-    );
-}
-
-#[test]
-fn add_renamed_table_fkey_ddl_sqlite() {
-    let (ops, new, ..) = create_add_renamed_table_fkey_ops();
-
-    let backend = butane_core::db::get_backend("sqlite").unwrap();
-    let sql = backend.create_migration_sql(&new, ops).unwrap();
-    let sql_lines: Vec<&str> = sql.lines().collect();
-    assert_eq!(
-        sql_lines,
-        vec![
-            "CREATE TABLE a_table (",
-            "id INTEGER NOT NULL PRIMARY KEY",
-            ");",
-            "CREATE TABLE b (",
-            "b INTEGER NOT NULL PRIMARY KEY,",
-            "FOREIGN KEY (b) REFERENCES a_table(id)",
-            ");",
-        ]
-    );
-}
-
-#[test]
-fn add_renamed_table_fkey_ddl_pg() {
-    let (ops, new, ..) = create_add_renamed_table_fkey_ops();
-
-    let backend = butane_core::db::get_backend("pg").unwrap();
-    let sql = backend.create_migration_sql(&new, ops).unwrap();
-    let sql_lines: Vec<&str> = sql.lines().collect();
-    assert_eq!(
-        sql_lines,
-        vec![
-            "CREATE TABLE a_table (",
-            "id INTEGER NOT NULL PRIMARY KEY",
-            ");",
-            "CREATE TABLE b (",
-            "b INTEGER NOT NULL PRIMARY KEY",
-            ");",
-            "ALTER TABLE b ADD FOREIGN KEY (b) REFERENCES a_table(id);",
-        ]
-    );
-}
-
-/// Creates the test case for adding a many table, returning the migration operations,
-/// the target ADB, and the tables which should be expected to be created.
-fn create_add_table_many_ops() -> (Vec<Operation>, ADB, ATable, ATable, ATable) {
-    let known_int_type = DeferredSqlType::KnownId(TypeIdentifier::Ty(SqlType::Int));
-
-    let old = ADB::default();
-    let mut new = ADB::default();
-
-    let id_column = AColumn::new(
-        "id".to_owned(),
-        known_int_type.clone(),
-        false, // nullable
-        true,  // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        None,  // reference
-    );
-
-    let mut table_a = ATable::new("a".to_owned());
-    table_a.add_column(id_column.clone());
-    new.replace_table(table_a.clone());
-
-    let mut table_b = ATable::new("b".to_owned());
-    table_b.add_column(id_column);
-
-    new.replace_table(table_b.clone());
-
-    let many_table = butane_core::migrations::adb::create_many_table(
-        "b",
-        "many_a",
-        DeferredSqlType::Deferred(TypeKey::PK("a".to_owned())),
-        "id",
-        known_int_type.clone(),
-    );
-    new.replace_table(many_table.clone());
-
-    new.resolve_types().unwrap();
-
-    let mut resolved_many_table = ATable::new(many_table.name);
-    let resolved_owner_column = AColumn::new(
-        "owner",
-        known_int_type.clone(),
-        false, // nullable
-        false, // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Literal(ARefLiteral::new(
-            "b".to_owned(),
-            "id".to_owned(),
-        ))),
-    );
-    let resolved_has_column = AColumn::new(
-        "has",
-        known_int_type.clone(),
-        false, // nullable
-        false, // pk
-        false, // auto
-        false, // unique
-        None,  // default
-        Some(ARef::Literal(ARefLiteral::new(
-            "a".to_owned(),
-            "id".to_owned(),
-        ))),
-    );
-    resolved_many_table.add_column(resolved_owner_column);
-    resolved_many_table.add_column(resolved_has_column);
-
-    let ops = diff(&old, &new);
-    (ops, new, table_a, table_b, resolved_many_table)
-}
-
-#[test]
-fn add_table_many() {
-    let (ops, _, table_a, table_b, resolved_many_table) = create_add_table_many_ops();
-
-    assert_eq!(ops[0], Operation::AddTable(table_a.clone()));
-    assert_eq!(ops[1], Operation::AddTable(table_b.clone()));
-    assert_eq!(ops[2], Operation::AddTable(resolved_many_table.clone()));
-
-    assert_eq!(
-        ops,
-        vec![
-            Operation::AddTable(table_a),
-            Operation::AddTable(table_b.clone()),
-            Operation::AddTable(resolved_many_table.clone()),
-            Operation::AddTableConstraints(resolved_many_table.clone()),
-        ]
-    );
-}
-
-#[test]
-fn add_table_many_ddl_sqlite() {
-    let (ops, new, ..) = create_add_table_many_ops();
-
-    let backend = butane_core::db::get_backend("sqlite").unwrap();
-    let sql = backend.create_migration_sql(&new, ops).unwrap();
-    let sql_lines: Vec<&str> = sql.lines().collect();
-    assert_eq!(
-        sql_lines,
-        vec![
-            "CREATE TABLE a (",
-            "id INTEGER NOT NULL PRIMARY KEY",
-            ");",
-            "CREATE TABLE b (",
-            "id INTEGER NOT NULL PRIMARY KEY",
-            ");",
-            "CREATE TABLE b_many_a_Many (",
-            "owner INTEGER NOT NULL,",
-            "has INTEGER NOT NULL,",
-            "FOREIGN KEY (owner) REFERENCES b(id)",
-            "FOREIGN KEY (has) REFERENCES a(id)",
-            ");",
-        ]
-    );
-}
-
-#[test]
-fn add_table_many_ddl_pg() {
-    let (ops, new, ..) = create_add_table_many_ops();
-
-    let backend = butane_core::db::get_backend("pg").unwrap();
-    let sql = backend.create_migration_sql(&new, ops).unwrap();
-    let sql_lines: Vec<&str> = sql.lines().collect();
-    assert_eq!(
-        sql_lines,
-        vec![
-            "CREATE TABLE a (",
-            "id INTEGER NOT NULL PRIMARY KEY",
-            ");",
-            "CREATE TABLE b (",
-            "id INTEGER NOT NULL PRIMARY KEY",
-            ");",
-            "CREATE TABLE b_many_a_Many (",
-            "owner INTEGER NOT NULL,",
-            "has INTEGER NOT NULL",
-            ");",
-            "ALTER TABLE b_many_a_Many ADD FOREIGN KEY (owner) REFERENCES b(id);",
-            "ALTER TABLE b_many_a_Many ADD FOREIGN KEY (has) REFERENCES a(id);",
-        ]
-    );
+    let to_apply = ms.unapplied_migrations(conn).unwrap();
+    assert_eq!(to_apply.len(), 2);
 }
